requirements.txt
attrs==24.3.0
certifi==2024.12.14
chardet==3.0.4
charset-normalizer==3.4.1
exceptiongroup==1.2.2
googletrans==4.0.0rc1
h11==0.9.0
h2==3.2.0
hpack==3.0.0
hstspreload==2025.1.1
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
idna==2.10
outcome==1.3.0.post0
packaging==24.2
PySocks==1.7.1
python-dotenv==1.0.1
requests==2.32.3
rfc3986==1.5.0
selenium==4.27.1
sniffio==1.3.1
sortedcontainers==2.4.0
tqdm==4.67.1
trio==0.27.0
trio-websocket==0.11.1
typing-extensions==4.12.2
urllib3==2.2.3
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
